# api/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Integracion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/etiquetas": {
            "get": {
                "tags": [
                    "Etiquetas"
                ],
                "summary": "Mostrar etiquetas",
                "operationId": "8a877e5ac008fd95bb7eebd08587864d",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los elementos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Etiquetas"
                ],
                "summary": "Crear etiqueta",
                "operationId": "17a1e627805adbab9080601668819b17",
                "responses": {
                    "200": {
                        "description": "Se ha creado la etiqueta."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/etiquetas/:id": {
            "get": {
                "tags": [
                    "Etiquetas"
                ],
                "summary": "Muestra la etiqueta por ID",
                "operationId": "258edad449ddcc018ba500216fc82da7",
                "responses": {
                    "200": {
                        "description": "Se ha completado la petición."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Etiquetas"
                ],
                "summary": "Actualizar etiquetas",
                "operationId": "6d203de4d34885f25a7101c745c1a351",
                "responses": {
                    "200": {
                        "description": "Se ha actualizado el elemento."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Etiquetas"
                ],
                "summary": "Borra etiquetas",
                "operationId": "b57002a9379baa09b1505d87a23253fe",
                "responses": {
                    "200": {
                        "description": "Se ha borrado el elemento."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer Token)",
                "name": "Authorization",
                "in": "header"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}